ht: fix file pseudo-request on Windows
The file:// URLs on windows have the form file:///c:/some/path with
three slashes which makes the Path of the URL /c:/some/path which
is not a valide file name on Windows.
Add a hack to strip this leading / on windows only.
No idea what happens with filenames of the form //host/share/path.