gui: new package for display and manipulation of data
The idea is to have a ht.Test in memory, display it in a HTML GUI which
allows to change the fields and execute the checks or even the request
and display the outcome. This GUI should guide the user writing proper
checks: This seems to be very hard as there are lots of tests and their
use and available fields is not present in memory. Learning all checks
by heart is hard; there is no "memory in the worls" to help remembering
their settings; there are no signifiers or constraints when writing
the checks as Hjson files. This makes writing checks very hard and
results in practice in tests without a full set of tests: Request works,
response looks good, check for status code 200, done.

The GUI should make it much easier and faster to write proper checks.

One of the main obstacles is the highly dynamic nature of the checks:
E.g. AnyOne is a check containing a list of checks or the JSON check
which may have an embedded JSON checks (which itself may have one
embdedded and so on). Thus a general mechanism is needed. This package
provides functions to render an (almost) arbitrary value type into a
HTML form and functions to update the value via the submitted form
parameters.