ht: fix unquoting for embedded JSON checks
Embedded JSON checks interpretes a string in a JSON document as an
embedded JSON document applicable to further checks. Such strings
must be unquoted properly to be a valid JSON document and just
dropping leading and trailing double quotes is not enough if the
embeded JSON document itself contains strings.

Fix this unquoting.