populate: fix panic when populating from null values
JSON null values converted to a soup with hjson and then fed into
populate resulted in a panic while trying to produce a nice error
message.
But there should be no error message at all: slices and maps can
be nil whcih is a natural mapping for a JSON null.