cmd/gui: register some types automatically
For Checks and Extratctors the main information is available: The type
and the Go doc and can be used to register the implements relation and
basic type info.
The field info is still combined in the doc of the type but it is better
than nothing and much faster than manual.