cmd/ht: improve help
Help for the subcommands was a bit odd:
Executing 'ht exec -help' would not print the command line flags, this was only
output during 'ht help exec'. Fix this and show flags always.
The usage information was for the subcommand only and the help message started
directly with this abbreviated usage. Make this nicer by moving the first line
of the help message before the usage and display the usage nicer and more copy-
pastable.