ht: fix error reporting on checks
What a shame. The Error field in a CheckResult is of type ErrorList to make
it explicit that this stores a set of errors. Unfortunately this set was
never really empty: For a passed test the error list contained one nil
entry making it non-empty. Dead stupid.
Must refactor this error list stuff properly.