gui: improve display of interface buttons
If you add a check in the GUI a list of buttons is displayed and each button
shows details of how to use this check as a tooltip. If this tooltip is
wider than the screen it is unreadable (and scrolling is basically impossible).
So group them in a div with a right margin.
While at it: Make tooltip even more wide and seperate buttons by breakable
spaces to prevent strange identations.