gui: add special case for multi-value properties
Several Checks contain string fields which contain flag controling the
check execution, e.g. ValidHTML has a Ignore field documeted as
    Ignore is a space separated list of issues to ignore.
and a long list of issues like 'lang', 'attr', 'url', ...
Such option fields are not the cleanes solution, these should be
individual boolean fields or maybe three-state toggles. But this
makes the Check struct unhandy large and ugly to read:
    IgnoreMalformdLangAttributes: true
In a GUI nobody expects such option fields to be a list of flags
in a text field.

This CL provides a new field setting which renders such fields
as a list of checkboxes.