gui: do not keep a copy on every update to a Value
Checkpointing the state of a Value should be an active process, e.g
there is a difference between adding a new empty/nil Check to a slice
and modifying a string value: The first is not worth as an undo-point.
So do not keep every value. This should also reduce memory consumption.