all: rename VarEx to DataExtraction in ht.Test
VarEx is cute but too hacky. And it's not really accurate because no variables
are extracted. The new tem should make it clearer that data is extracted from
the response. These extractions are 'named': The key in the map is the name.
This name is used to update the corresponding variable in the global scope
after executing the Test.