docker: use a multistage build for the final ht image
Multistage builds make it much easier to drop unused layers in the
final image. With this the whole Docker image build process is now
decoupled from my laptop.
TLS certs are needed for https and the MySQL driver (probably)
needs the zoneinfo.

The build context must be the root dir of this project but I'd like
to keep the docker stuff in a subfolder.