ht: handle escaping properly in ValidHTML check
Unescaped & are bad but HTML5 allows lots of unescaped &. In attributes it
is quite hard to produce something invalid: The only invalid stuff are
unknown character references like '&GhiPoXs;'. As & are very common in
attributes the ValidHTML check was not very useful. The first fix for this
problem was made in c38b623ca: A new issue to ignore ('attresc') was
introduced which ignored escaping problems in attributes. This is plain
awful: You have to ignore a nonexisting problem to make the check pass
because the check is broken.

This CL hopefuly does a better job: Checking escaping can be turned of
via 'escaping' but only as a global option, the 'attrescp' is gone.
Escpaing checking in attributes is now done properly: Only ambiguous
ampersands (i.e. ampersands which might start a misspelled entity) are
reported.