ht: new extractors
Add a SetTimestamp extractor to allow generation of current
timestamps in various formats.

Also: eexperiment with arithemtic expression eval in SetVariable
This might come handy if you do not have much control over the test data,
e.g. in a smoke test on a production environment: This would allow you
to look up any product, extract its price and check that the total cart
sum actually matches '2*{{PRICE}} * 1.08' if you add two pieces of the
product to the cart and add 8% VAT.

Needs vendoring of govaluate.