suite: indicate clipped HTTP bodies in HTML report
If the response body gets 'too big' it is clipped. Long lines are clipped
at the end and if too many lines a present the middle lines are clipped.
These clippings are marked with horizontal (U+2026) and vertical ellipsises
(U+22EE). Unfortunatly this elipsises are easy to overlook. This tripped
users into beliving that the actual response was buggy as it was missing
several lines.

This change highlights any clipped body and adds a title attribute to
make it clear: 'This is _not_ the original response but just a preview
for your convenience; please klick the link below which takes you to
the real response.' Hope this helps prevent user confusion.