gui: fix turning off boolean flag
Bools are rendered as checkboxes and a unticked checkbox does not produce a
value. We thus never could turn off a boolean flag i.e. set it to false once
it was ticked.

Change the mechanism for bools to "set to true if some value is sent and
set to false if no parameter is sent".