ht: don't evaluate expression in SetVariable extractor
The change introduced in commit 409b4fe allowed to evaluate an expression and
set a variable to the evaluated value. This was an experiment. Unfurtunately
it is a stupid idea: We already do have an extractor which allows to evaluate
expressions, the JSExtractor. This extractor works better: It can access
slice and map elements, has more complete string functions and its syntax
is by far more common.

So undo the experiment.