cmd/ht: add example subcommand
The idea is to have a hierarchy of examples and the user can browse down
this hierarchy. A bit like gnuplots builtin documentation. This should
solve the problem that it is hard to come up with a scratch Test or Suite
by just reading the documentation.

The examples are validated during go test so the should not fall behind
the actual code (like the showcase did several times)

Remove examples from the help subcommand as they are superseeded by the
new example subcommand

Add go:generate directives for generated stuff, not only for the examples
but for the doc and gui subcommands too.

Still to do: Write more examples, write more example suites and write
an example load test. Maybe some examples of how to work in the commandline
too.