cmd/ht: add example subcommand
The idea is to have a hierarchy of examples and the user can browse down
this hierarchy. A bit like gnuplots builtin documentation. This should
solve the problem that it is hard to come up with a scratch Test or Suite
by just reading the documentation.

Still missing: pregeneration of the examples to bake into the executable
and automatic validation that the examples are functional.