ht: fix line number in ValidHTML errors
The reported line numbers were wrong: Line numbers were increased not only
for newlines but also if the full io.Reader buffer was used (which happens on
long lines and all lines in the unit test were short). Major embarasing fuckup.
Fix this.
With long lines it also got obvious that reporting only the line number makes
debugging hard: Lines of several thousend characters are common in generated
or compacted HTML and finding a mismatched tag or an unescaped > is very
painful. Report the colum where the error got detected. This is always too
late but better than nothing.
Reporting the 'global' errors after the last line is not realy intuitive.
This CL removes reporting a line number for this global errors like duplicate
ids because the line number reported is not helpful at all. While at it: Make
the error for DOCTYPE clearer.

Also add more tests.