cmd/ht: provide current working dir in CWD variable
It is possible to get the path of the current suite and the current test
through the SUITE_DIR and TEST_Dir variables but these are just relative
pathsto the current working dir. This is fine for most purposes like
reading files with the @file syntax (e.g. a request body). It works
because relative paths work pretty well.

Unfortunately the pseudo requests encode paths in the path of the URL
and URL paths are not relative, they are always rooted as the root /
separates the host from the path.
Most probably this is not a big problem in a productive use of ht but
it is painful in testing e.g. the file:// pseudo requests. On Unix-like
systems you can rely on /etc/passwd beeing present but it gets complicated
for writing /tmp might not be available (or linked somewhere) on OSX.

This change automatically adds the variable CWD (if not set via the
command line) to the current working directory. The name is common
and too short to be used as a 'real' variable name so it should be okay.
The possible alternative to this solution during test is used already
in use: Find the TEMDIR or the CWD and manualy inject into the the test.
But this alternative is not good for the examples: The examples should
be directly usable, without a "don't forget to manually set CWD on
the command line or replace its use with an absolute path".

The CWD variable is set by cmd/ht, not in package suite or package ht
which means you can still work with package suite or ht as usual.