ht: rework Latency check, improve and fix it
The Latency check's implementation was a candidate for a git-filter-branch
complete history rewrite. It was broken. Logically and technically. E.g.
it might hang indefinitively because goroutines where blocking them without
all becoming asleep. It's data collection code was plain stupid.
Handling of failed or errored tests during the latency check where not
handled in a sensible manner. The tests where lacking. What an embarrassing
mess.

This is almost a rewrite, it gets the collection logic straight so it
does no longer hang and handles error/failures as a failure of the latency
check ("The latency of this application is good, we meet our SLA goals:
We server 500 Internal Server Error in under 6ms!" is not what we want.)
Refactor into smaller parts and fix tests and add one more test.