mock: make cookies extractable from request
It might make sense to read a cookie sent in the request and craft a
dynamic response based on the cookie value. Unfortunately the standard
ht.CookieExtractor works on Set-Cookie headers of a response and not the
Cookie header of a request. To make the standard cookie extractor work
we rewrite Cookie to Set-Cookie headers.