ht: improve error message for failed extractors
Found during debugging a mock. The error message was not clear at all.
Quoting the unrecognised named would have helped a bit, but it would clutter
the normal error message when someone misspelled the extractor and not the
Extractor key.
While at it: remove the 'ht: ' prefix: The are anoying when using cmd/ht.
Any program which uses ht as a package could augment the error by itself.