ht: improve csv export of Latency check
Skip doing unnecessary work if the underlying writer would discard the
work anyway. Document the exported columns (at least a bit) and add a
column for the thread which executed this request. Change the order of
the columns from general to specific with the most interesting column
(the response time) going last.
Formaly these changes might break external tools which rely on the
columns: The fix should be simple and it would break anyway sometimes,
so why not now.