cmd/ht: expose default client timeout on command line
The package ht exposes the default HTTP client timeout but this was not
exposed in cmd/ht. The package default of 10s is pretty sensible but might
be awkward during testing: If you have to work with very slow backend
systems the 10s might be too slow but changing all Tests to use a custom
timeout is also a lot of work. Thus expose the default on the command line.