cmd/ht: improve tests of cmd/ht
We had a few tests of the binary, namely the cmd/ht/testdata/test-*.bash
files which check that the ht command behaves properly. This CL adds
a test for the exit code of the two subcommands 'run' end 'exec'.

Keeping these tests ander testdata is a bit wrong so move them to
cmd/ht directly.