suite: use sample size dependent tolerances when checking distribution
The original 5% tolerance for contributions of a scenario to the total
oad generated in a throughput test results in lots of failing tests:
If only 100 requests are measured at all and 20% should come from one
scenario then it is not unlikely that e.g. 27 were drawn from this
scenario. This are formaly 7 percent points error but this is statistically
expectable.
Use tolerances depending on the sample size: For larger sample sizes
the variation should become less but for short load tests accept larger
variations in actual scenario distribution.