all: introduce ht logger type and use it
Several packages need to log stuff in various situations. Some used their
own logger interface, some just logged to stdout (notably the throughput
tests in package suite).
This CL introduces an exported Logger type in ht and uses this one in
the ht, mock and suite packages.

Formaly this is a braking API change as this logger has to be injected
into exported APIs but I tend to see this as a bugfix with logging to
stdout being the bug.