ht: allow to check for weak ETags
Our ETag check checks that the tag is present and that a subsequent GET
with that etag in a If-None-Match header results in a 304 Not Modified.
Up to know only strict ETags could be checked like this, but the actual
check is totaly the same for weak ETags too. ("Weak" here is "only
semanticaly equivalent" while "stong" is "byte-by-byte identical").

So make the ETag check to check weak ETags too. As a weak ETag is, well,
weaker than a strong ETag this is an option which must be enabled in the
check.