cmd/ht add -show cmdline flag to open browser
During an interactive session (be it test creation, test debugging or
explorative testing) it is cumpersome to copy/paste the URL of the generated
output to the browsers. Of course it is possible to use -output and a fixed
directory and keep the report which is regenerated in the browser and refresh
it but this overwrites previous test runs which one might want to keep.

This CL adds a new command line flag -show to the commands quick, run and exec
which opens a webbrowser with the generated report. For multi-suite execs it
opens the overall report once and the user can refresh it in the browser
after each suite manually. This seems the best solution: The results are
not overwritten as in the fixed -output variant described above and opening
each individual suite report is wasteful.

Tested on Mac only, but should work on Windows and Linux too.