Restructure handling of pseudo-request
In the moment we have 3 pseudo-requests versus 1 real HTTP(S) request.
And it seems natural to add a Kafka, Redis, MongoDB, ZeroMQ, etc.
request too. Ultimately this might lead to replacing the Request
with a generic Action which would have the nice effect that we no
longer would need to cramp configurations for these actions
(like the SQL data source name) into magic query parameters or
header fields. For now limit the refactoring to something
simple.
What needs either to be changed or to be justified is the use
of an interface for the actions with empty structs implementing
it (which shows that a struct with functions would do too).